+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- prevent adding AUTHORED component to an IMPORTED component using require relative syntax
 - avoid writing long files paths for imported components when there is a shared directory among the component files and its dependencies
 
 ## [0.11.1] - 2017-11-29